requirements.txt
langchain
tiktoken
qdrant-client
torch
transformers
accelerate
bitsandbytes
sentence_transformers
streamlit
python-dateutil
openai
tqdm